Convert a bunch of event-handling code to TS; move into content app
