+++ content/webapp/package.json
@@ -27,6 +27,7 @@
     "koa": "^2.5.0",
     "koa-bodyparser": "^4.2.1",
     "koa-router": "^7.4.0",
+    "lodash.sortby": "^4.7.0",
     "lodash.throttle": "^4.1.1",
     "next": "^11.1.2",
     "node-fetch": "^2.6.2",
