+++ content/webapp/pages/homepage.tsx
@@ -15,11 +15,11 @@ import Layout10 from '@weco/common/views/components/Layout10/Layout10';
 import SimpleCardGrid from '../components/SimpleCardGrid/SimpleCardGrid';
 import PageHeaderStandfirst from '../components/PageHeaderStandfirst/PageHeaderStandfirst';
 import { getExhibitions } from '@weco/common/services/prismic/exhibitions';
+import { getEvents } from '@weco/common/services/prismic/events';
 import {
-  getEvents,
   orderEventsByNextAvailableDate,
   filterEventsForNext7Days,
-} from '@weco/common/services/prismic/events';
+} from '../services/prismic/events';
 import { UiExhibition } from '@weco/common/model/exhibitions';
 import { UiEvent } from '@weco/common/model/events';
 import { convertJsonToDates } from './event';
